# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /admin/loans:
    get:
      operationId: getAdminLoans
      summary: List all active loans
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
  /admin/loans/{id}:
    get:
      operationId: getAdminLoan
      summary: Get a single loan with repayments and penalties
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanDetail"
    patch:
      operationId: updateAdminLoan
      summary: Update loan status/notes and optionally send SMS
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /admin/loans/{id}/penalty:
    post:
      operationId: addLoanPenalty
      summary: Add penalty to a loan
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PenaltyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Penalty"
  /admin/loans/{id}/repayment:
    post:
      operationId: recordRepayment
      summary: Record a loan repayment
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repayment"
  /admin/loans/{id}/sms:
    post:
      operationId: sendLoanSms
      summary: Send SMS notification to loan client
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"

  # ── ADMIN: OVERDUES ───────────────────────────────────
...
  /admin/notifications:
    get:
      operationId: getAdminNotifications
      summary: List all notifications sent
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: sendNotification
      summary: Send a notification/SMS to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── ADMIN: USERS ──────────────────────────────────────
  /admin/users:
    get:
      operationId: getAdminUsers
      summary: List all applicant users
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicantUser"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      summary: Get single user with full history
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicantUserDetail"

  # ── ADMIN: AUDIT LOG ─────────────────────────────────
...
components:
  schemas:
...
    OkResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]

    # ── APPLICATION ──────────────────────────────────────
...
    Application:
      type: object
      required: [id, fullName, nationalId, phone, email, address, employmentType, employer, monthlyIncome, requestedAmount, repaymentMonths, payoutMethod, status, referenceNumber, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        nationalId:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        employmentType:
          type: string
        employer:
          type: string
        monthlyIncome:
          type: number
        requestedAmount:
          type: number
        repaymentMonths:
          type: integer
        payoutMethod:
          type: string
        status:
          type: string
        referenceNumber:
          type: string
        adminNotes:
          type: ["string", "null"]
        approvedAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Loan:
      type: object
      required: [id, applicationId, principalAmount, interestRate, repaymentMonths, monthlyPayment, totalRepayable, outstandingBalance, totalPaid, penaltyAmount, status, disbursedAt, dueDate, createdAt, updatedAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        applicantName:
          type: ["string", "null"]
        applicantPhone:
          type: ["string", "null"]
        applicantEmail:
          type: ["string", "null"]
        principalAmount:
          type: number
        interestRate:
          type: number
        repaymentMonths:
          type: integer
        monthlyPayment:
          type: number
        totalRepayable:
          type: number
        outstandingBalance:
          type: number
        totalPaid:
          type: number
        penaltyAmount:
          type: number
        status:
          type: string
        payoutMethod:
          type: ["string", "null"]
        disbursedAt:
          type: string
        dueDate:
          type: string
        nextDueDate:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LoanUpdate:
      type: object
      properties:
        status:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        nextDueDate:
          type: ["string", "null"]
        sendSms:
          type: ["boolean", "null"]
        smsMessage:
          type: ["string", "null"]
    LoanDetail:
      allOf:
        - $ref: "#/components/schemas/Loan"
        - type: object
          properties:
            repayments:
              type: array
              items:
                $ref: "#/components/schemas/Repayment"
            penalties:
              type: array
              items:
                $ref: "#/components/schemas/Penalty"
...
    PenaltyInput:
      type: object
      required: [amount, reason]
      properties:
        amount:
          type: number
        reason:
          type: string
    Penalty:
      type: object
      required: [id, loanId, amount, reason, createdAt]
      properties:
        id:
          type: integer
        loanId:
          type: integer
        amount:
          type: number
        reason:
          type: string
        createdAt:
          type: string
    RepaymentInput:
      type: object
      required: [amount, paymentMethod]
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        reference:
          type: ["string", "null"]
    Repayment:
      type: object
      required: [id, loanId, amount, paymentMethod, paidAt]
      properties:
        id:
          type: integer
        loanId:
          type: integer
        amount:
          type: number
        paymentMethod:
          type: string
        reference:
          type: ["string", "null"]
        paidAt:
          type: string
    SmsInput:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── NOTIFICATIONS ────────────────────────────────────
    NotificationInput:
      type: object
      required: [recipientPhone, message, type]
      properties:
        recipientPhone:
          type: string
        recipientEmail:
          type: ["string", "null"]
        message:
          type: string
        type:
          type: string
        applicationId:
          type: ["integer", "null"]
        loanId:
          type: ["integer", "null"]
    Notification:
      type: object
      required: [id, recipientPhone, message, type, status, createdAt]
      properties:
        id:
          type: integer
        recipientPhone:
          type: string
        recipientEmail:
          type: ["string", "null"]
        message:
          type: string
        type:
          type: string
        status:
          type: string
        applicationId:
          type: ["integer", "null"]
        loanId:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ── USERS ────────────────────────────────────────────
    ApplicantUser:
      type: object
      required: [id, fullName, email, phone, nationalId, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        nationalId:
          type: string
        applicationCount:
          type: ["integer", "null"]
        createdAt:
          type: string
    ApplicantUserDetail:
      allOf:
        - $ref: "#/components/schemas/ApplicantUser"
        - type: object
          properties:
            applications:
              type: array
              items:
                $ref: "#/components/schemas/Application"

    # ── AUDIT LOG ────────────────────────────────────────
...